Python/requirements.txt
Pillow==2.6.1
numpy==1.9.1
wsgiref==0.1.2
